Refaactor task rules to RuleMixin
