+++ griptape/mixins/__init__.py
@@ -1,5 +1,6 @@
 from .activity_mixin import ActivityMixin
 from .exponential_backoff_mixin import ExponentialBackoffMixin
 from .action_subtask_origin_mixin import ActionSubtaskOriginMixin
+from .rule_mixin import RuleMixin
 
-__all__ = ["ActivityMixin", "ExponentialBackoffMixin", "ActionSubtaskOriginMixin"]
+__all__ = ["ActivityMixin", "ExponentialBackoffMixin", "ActionSubtaskOriginMixin", "RuleMixin"]
