+++ griptape/tasks/text_to_image_task.py
@@ -2,16 +2,16 @@ from __future__ import annotations
 
 from typing import Callable
 
-from attr import define
+from attr import define, field
 
 from griptape.artifacts import ImageArtifact, TextArtifact
-from griptape.tasks import BaseImageGenerationTask
+from griptape.tasks import BaseImageGenerationTask, BaseTask
 from griptape.utils import J2
 
 
 @define
 class TextToImageTask(BaseImageGenerationTask):
-    """ImageGenerationTask is a task that can be used to generate an image. Accepts a text prompt as input in one of
+    """TextToImageTask is a task that can be used to generate an image. Accepts a text prompt as input in one of
     the following formats:
     - template string
     - TextArtifact
@@ -24,6 +24,10 @@ class TextToImageTask(BaseImageGenerationTask):
         output_dir: If provided, the generated image will be written to disk in output_dir.
         output_file: If provided, the generated image will be written to disk as output_file.
     """
+
+    DEFAULT_INPUT_TEMPLATE = "{{ args[0] }}"
+
+    _input: str | TextArtifact | Callable[[BaseTask], TextArtifact] = field(default=DEFAULT_INPUT_TEMPLATE)
 
     @property
     def input(self) -> TextArtifact:
