+++ tests/mocks/mock_image_generation_task.py
@@ -6,5 +6,7 @@ from griptape.tasks import BaseImageGenerationTask
 
 @define
 class MockImageGenerationTask(BaseImageGenerationTask):
+    input: str = field(default="input")  # pyright: ignore
+
     def run(self) -> ImageArtifact:
         return ImageArtifact(value=b"image data", mime_type="image/png", width=512, height=512)
